+++ src/R-yyjson-parse.c
@@ -6,6 +6,7 @@
 #include <Rdefines.h>
 
 #include <stdio.h>
+#include <stdbool.h>
 #include <stdlib.h>
 #include <unistd.h>