+++ src/geojson-serialize.c
@@ -6,6 +6,7 @@
 #include <Rdefines.h>
 
 #include <stdio.h>
+#include <stdbool.h>
 #include <stdlib.h>
 #include <unistd.h>
 #include <time.h>
@@ -112,13 +113,16 @@ yyjson_mut_val *serialize_geom(SEXP sf_, yyjson_mut_doc *doc, geo_serialize_opti
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP sfc_to_str(SEXP sfc_, geo_serialize_options *opt) {
+SEXP sfc_to_str(SEXP sfc_, geo_serialize_options *opt, bool as_raw) {
   
   int nprotect = 0;
   if (!Rf_isNewList(sfc_)) {
-    Rf_error("serialize_sfc(): Expeting list\n");
+    Rf_error("sfc_to_str(): Expeting list\n");
   }
   
+  if (as_raw) {
+    Rf_error("sfc_to_str(): Not currently possible to serialize 'sfc' to raw vector");
+  }
   
   R_xlen_t N = Rf_xlength(sfc_);
   SEXP geojson_ = PROTECT(Rf_allocVector(STRSXP, N)); nprotect++;
@@ -225,7 +229,7 @@ yyjson_mut_doc *sf_to_json(SEXP sf_, geo_serialize_options *opt) {
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // sf data.frame to feature collection
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP sf_to_str(SEXP sf_, geo_serialize_options *opt) {
+SEXP sf_to_str(SEXP sf_, geo_serialize_options *opt, bool as_raw) {
   
   int nprotect = 0;
   yyjson_mut_doc *doc = sf_to_json(sf_, opt);
@@ -243,7 +247,14 @@ SEXP sf_to_str(SEXP sf_, geo_serialize_options *opt) {
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   // Convert string to R character, tidy and return
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  SEXP geojson_ = PROTECT(Rf_mkString(json)); nprotect++;
+  SEXP geojson_;
+  if (as_raw) {
+    geojson_ = PROTECT(Rf_allocVector(RAWSXP, strlen(json))); nprotect++;
+    memcpy(RAW(geojson_), json, strlen(json));
+  } else {
+    geojson_ = PROTECT(Rf_mkString(json)); nprotect++;
+  }
+  
   free(json);
   yyjson_mut_doc_free(doc);
   UNPROTECT(nprotect);
@@ -285,7 +296,7 @@ SEXP sf_to_file(SEXP sf_, SEXP filename_, geo_serialize_options *opt) {
 //
 // Serialize R object to JSON string.  Callable from R
 //===========================================================================
-SEXP serialize_sf_to_str_(SEXP sf_, SEXP to_geo_opts_, SEXP serialize_opts_) {
+SEXP serialize_sf_to_str_(SEXP sf_, SEXP to_geo_opts_, SEXP serialize_opts_, SEXP as_raw_) {
 
   int nprotect = 0;
   if (!Rf_inherits(sf_, "sf") && !Rf_inherits(sf_, "sfc")) {
@@ -298,9 +309,9 @@ SEXP serialize_sf_to_str_(SEXP sf_, SEXP to_geo_opts_, SEXP serialize_opts_) {
 
   SEXP res_ = R_NilValue;
   if (Rf_inherits(sf_, "sfc")) {
-    res_ = PROTECT(sfc_to_str(sf_, &opt)); nprotect++;
+    res_ = PROTECT(sfc_to_str(sf_, &opt, Rf_asLogical(as_raw_))); nprotect++;
   } else if (Rf_inherits(sf_, "sf")) {
-    res_ = PROTECT(sf_to_str(sf_, &opt)); nprotect++;
+    res_ = PROTECT(sf_to_str(sf_, &opt, Rf_asLogical(as_raw_))); nprotect++;
   } else {
     Rf_error("serialize_sf_to_str_: class not handled yet");
   }
@@ -308,6 +319,7 @@ SEXP serialize_sf_to_str_(SEXP sf_, SEXP to_geo_opts_, SEXP serialize_opts_) {
   UNPROTECT(nprotect);
   return res_;
 }
+
 
 SEXP serialize_sf_to_file_(SEXP sf_, SEXP filename_, SEXP to_geo_opts_, SEXP serialize_opts_) {