+++ src/init.c
@@ -14,7 +14,7 @@ extern SEXP parse_from_str_ (SEXP str_     , SEXP parse_opts_);
 extern SEXP parse_from_file_(SEXP filename_, SEXP parse_opts_);
 extern SEXP parse_from_raw_ (SEXP filename_, SEXP parse_opts_);
 
-extern SEXP serialize_to_str_ (SEXP x_,                 SEXP serialize_opts_);
+extern SEXP serialize_to_str_ (SEXP x_,                 SEXP serialize_opts_, SEXP as_raw_);
 extern SEXP serialize_to_file_(SEXP x_, SEXP filename_, SEXP serialize_opts_);
 
 extern SEXP validate_json_file_(SEXP filename_, SEXP verbose_, SEXP parse_opts_);
@@ -29,10 +29,10 @@ extern SEXP parse_ndjson_file_as_list_(SEXP filename_, SEXP nread_, SEXP nskip_,
 extern SEXP parse_ndjson_str_as_df_  (SEXP str_, SEXP nread_, SEXP nskip_, SEXP nprobe_, SEXP parse_opts_);
 extern SEXP parse_ndjson_str_as_list_(SEXP str_, SEXP nread_, SEXP nskip_,               SEXP parse_opts_);
 
-extern SEXP serialize_df_to_ndjson_str_ (SEXP robj_,                 SEXP serialize_opts_);
+extern SEXP serialize_df_to_ndjson_str_ (SEXP robj_,                 SEXP serialize_opts_, SEXP as_raw_);
 extern SEXP serialize_df_to_ndjson_file_(SEXP robj_, SEXP filename_, SEXP serialize_opts_);
 
-extern SEXP serialize_list_to_ndjson_str_ (SEXP robj_,                 SEXP serialize_opts_);
+extern SEXP serialize_list_to_ndjson_str_ (SEXP robj_,                 SEXP serialize_opts_, SEXP as_raw_);
 extern SEXP serialize_list_to_ndjson_file_(SEXP robj_, SEXP filename_, SEXP serialize_opts_);
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -41,7 +41,7 @@ extern SEXP serialize_list_to_ndjson_file_(SEXP robj_, SEXP filename_, SEXP seri
 extern SEXP parse_geojson_str_ (SEXP str_     , SEXP geo_opts_, SEXP parse_opts_);
 extern SEXP parse_geojson_file_(SEXP filename_, SEXP geo_opts_, SEXP parse_opts_);
 
-extern SEXP serialize_sf_to_str_ (SEXP sf_                , SEXP geo_opts_, SEXP serialize_opts_);
+extern SEXP serialize_sf_to_str_ (SEXP sf_                , SEXP geo_opts_, SEXP serialize_opts_, SEXP as_raw_);
 extern SEXP serialize_sf_to_file_(SEXP sf_, SEXP filename_, SEXP geo_opts_, SEXP serialize_opts_);
 
 
@@ -52,7 +52,7 @@ static const R_CallMethodDef CEntries[] = {
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   // Regular JSON
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  {"serialize_to_str_" , (DL_FUNC) &serialize_to_str_ , 2},
+  {"serialize_to_str_" , (DL_FUNC) &serialize_to_str_ , 3},
   {"serialize_to_file_", (DL_FUNC) &serialize_to_file_, 3},
   
   {"parse_from_str_"  , (DL_FUNC) &parse_from_str_ , 2},
@@ -71,10 +71,10 @@ static const R_CallMethodDef CEntries[] = {
   {"parse_ndjson_str_as_df_"  , (DL_FUNC) &parse_ndjson_str_as_df_  , 5},
   {"parse_ndjson_str_as_list_", (DL_FUNC) &parse_ndjson_str_as_list_, 4},
   
-  {"serialize_df_to_ndjson_str_" , (DL_FUNC) &serialize_df_to_ndjson_str_ , 2},
+  {"serialize_df_to_ndjson_str_" , (DL_FUNC) &serialize_df_to_ndjson_str_ , 3},
   {"serialize_df_to_ndjson_file_", (DL_FUNC) &serialize_df_to_ndjson_file_, 3},
   
-  {"serialize_list_to_ndjson_str_" , (DL_FUNC) &serialize_list_to_ndjson_str_ , 2},
+  {"serialize_list_to_ndjson_str_" , (DL_FUNC) &serialize_list_to_ndjson_str_ , 3},
   {"serialize_list_to_ndjson_file_", (DL_FUNC) &serialize_list_to_ndjson_file_, 3},
   
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -83,7 +83,7 @@ static const R_CallMethodDef CEntries[] = {
   {"parse_geojson_str_" , (DL_FUNC) &parse_geojson_str_ , 3},
   {"parse_geojson_file_", (DL_FUNC) &parse_geojson_file_, 3},
 
-  {"serialize_sf_to_str_" , (DL_FUNC) &serialize_sf_to_str_ , 3},
+  {"serialize_sf_to_str_" , (DL_FUNC) &serialize_sf_to_str_ , 4},
   {"serialize_sf_to_file_", (DL_FUNC) &serialize_sf_to_file_, 4},