+++ tests/testthat/test-write-json-raw.R
@@ -1,0 +1,13 @@
+
+test_that("write_json_raw() works", {
+
+  dat <- head(iris, 3)
+  dat$Species <- as.character(dat$Species)
+  
+  js <- write_json_raw(dat)
+  expect_true(is.raw(js))
+  res <- read_json_raw(js)
+  expect_identical(res, dat)
+  
+  
+})